requirements.txt
SoundFile
torchaudio
torch
torchvision
munch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
opencv-python
Pillow
face_alignment
hydra-core
omegaconf
decord
ffmpeg-python
imageio
safetensors
modelscope
huggingface_hub
phonemizer
scipy
pandas
click
numpy
iopath
pycocotools
requests
audio-separator
openai
faster-whisper
git+https://github.com/resemble-ai/monotonic_align.git